Reorganize the file and add small comments.
